api/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
motor==3.0.0
dnspython==2.0.0
email-validator==1.3.0
pydantic==1.10.2
pyflakes==2.2.0
pymongo==4.2.0
starlette==0.20.4
bcrypt==3.2.0
dnspython==2.0.0
flake8==3.8.4
passlib[bcrypt]==1.7.4
python-decouple==3.6
python-multipart==0.0.5
openpyxl==3.0.10
pyexcel==0.7.0
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==2.5.0
python-jose[cryptography]==3.3.0
fastapi-mail==1.1.5
pytz==2022.4